Move layer functions to utils.py
